# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the signed-in profile and staff role
      responses:
        "200":
          description: Active signed-in session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401": { description: Authentication required }
...
components:
...
  schemas:
...
    AuthSession:
      type: object
      required: [profile]
      properties:
        profile:
          type: object
          required: [id, firstName, lastName, email, emailVerified, phoneVerified, role]
          properties:
            id: { type: string }
            firstName: { type: string }
            lastName: { type: string }
            email: { type: string }
            phone: { type: string, nullable: true }
            emailVerified: { type: boolean }
            phoneVerified: { type: boolean }
            role: { type: string, enum: [customer, driver, admin, dispatcher, support] }
...